Move whitelist to separate directory
